+++ core/src/main/java/com/cobber/fta/dates/SimpleDateMatcher.java
@@ -129,7 +129,7 @@ public class SimpleDateMatcher {
 		matchers.add(new SimpleFacts("EEEE, MMM, d{2}, d{4}", "EEEE, MMM, dd, yyyy", FTAType.LOCALDATE));
 		matchers.add(new SimpleFacts("EEEE, MMMM, d{2}, d{4}", "EEEE, MMMM, dd, yyyy", FTAType.LOCALDATE));
 		matchers.add(new SimpleFacts("EEE MMM d{2} d{2}:d{2}:d{2} z d{4}", "EEE MMM dd HH:mm:ss z yyyy", FTAType.ZONEDDATETIME));
-		matchers.add(new SimpleFacts("EEE MMM {2}d d{2}:d{2}:d{2} z d{4}", "EEE MMM ppd HH:mm:ss z yyyy", FTAType.ZONEDDATETIME));
+		matchers.add(new SimpleFacts("EEE MMM  d d{2}:d{2}:d{2} z d{4}", "EEE MMM ppd HH:mm:ss z yyyy", FTAType.ZONEDDATETIME));
 		matchers.add(new SimpleFacts("EEE MMM d{2} d{2}:d{2}:d{2} O d{4}", "EEE MMM dd HH:mm:ss O yyyy", FTAType.OFFSETDATETIME));
 		matchers.add(new SimpleFacts("EEE MMM d{2} d{2}:d{2}:d{2} OOOO d{4}", "EEE MMM dd HH:mm:ss OOOO yyyy", FTAType.OFFSETDATETIME));
 
@@ -311,18 +311,6 @@ public class SimpleDateMatcher {
 					lastCh = ch;
 				}
 				result.append(ch);
-			}
-			else  if (ch == ' ') {
-				if (lastCh == ' ')
-					count++;
-				else {
-					if (count != 0) {
-						result.append('{').append(count + 1).append('}');
-						count = 0;
-					}
-					result.append(' ');
-					lastCh = ch;
-				}
 			}
 			else {
 				if (count != 0) {